+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/models/flood/DeterministicFloodHazard.java
@@ -1,0 +1,22 @@
+/*******************************************************************************
+ * Copyright (c) 2020 University of Illinois and others.  All rights reserved.
+ * This program and the accompanying materials are made available under the
+ * terms of the Mozilla Public License v2.0 which accompanies this distribution,
+ * and is available at https://www.mozilla.org/en-US/MPL/2.0/
+ *
+ * Contributors:
+ *******************************************************************************/
+package edu.illinois.ncsa.incore.service.hazard.models.flood;
+
+public class DeterministicFloodHazard extends FloodHazardDataset {
+    private FloodDeterministicParameters floodParameters;
+
+    public FloodDeterministicParameters getFloodParameters() {
+        return floodParameters;
+    }
+
+    public void setFloodParameters(FloodDeterministicParameters floodParameters) {
+        this.floodParameters = floodParameters;
+    }
+
+}